Move Value and Closeable implementations to base class
