+++ src/main/php/util/address/Address.class.php
@@ -1,16 +1,30 @@
 <?php namespace util\address;
 
 use Iterator, IteratorAggregate, Traversable;
-use lang\IllegalStateException;
-use util\NoSuchElementException;
+use io\Channel;
+use io\streams\{InputStream, MemoryInputStream};
+use lang\{Closeable, Value};
+use util\{NoSuchElementException, Objects};
 
 /**
  * Base class for all XML inputs
  *
  * @test  xp://util.address.unittest.AddressTest
  */
-abstract class Address implements IteratorAggregate {
+abstract class Address implements Closeable, Value, IteratorAggregate {
   private $iterator;
+  protected $stream;
+
+  /** @param string|io.Channel|io.streams.InputStream $source */
+  public function __construct($source) {
+    if ($source instanceof InputStream) {
+      $this->stream= $source;
+    } else if ($source instanceof Channel) {
+      $this->stream= $source->in();
+    } else {
+      $this->stream= new MemoryInputStream($source);
+    }
+  }
 
   /**
    * Returns a stream
@@ -18,9 +32,7 @@ abstract class Address implements IteratorAggregate {
    * @deprecated Implement iterator() instead!
    * @return io.streams.InputStream
    */
-  protected function stream() {
-    throw new IllegalStateException('Implement either stream() or iterator() in subclasses');
-  }
+  protected function stream() { return $this->stream; }
 
   /**
    * Creates an iterator. Default implementation is to return an
@@ -124,4 +136,34 @@ abstract class Address implements IteratorAggregate {
 
     throw new NoSuchElementException('No more elements in iterator');
   }
+
+  /** @return string */
+  public function toString() {
+
+    // Most InputStream instances have a toString() method but do not implement
+    // the Value interface, see https://github.com/xp-framework/core/issues/310
+    if ($this->stream instanceof Value || method_exists($this->stream, 'toString')) {
+      return nameof($this).'<'.$this->stream->toString().'>';
+    } else {
+      return nameof($this).'<'.Objects::stringOf($this->stream).'>';
+    }
+  }
+
+  /** @return string */
+  public function hashCode() {
+    return 'XS'.Objects::hashOf($this->stream);
+  }
+
+  /**
+   * Comparison
+   *
+   * @param  var $value
+   * @return int
+   */
+  public function compareTo($value) {
+    return $value instanceof self ? Objects::compare($this->stream, $value->stream) : 1;
+  }
+
+  /** @return void */
+  public function close() { $this->stream->close(); }
 }
