+++ src/main/php/util/address/XmlFile.class.php
@@ -1,7 +1,7 @@
 <?php namespace util\address;
 
-use io\File;
-use lang\{Closeable, Value};
+use Iterator;
+use io\streams\FileInputStream;
 
 /**
  * XML file input
@@ -9,8 +9,7 @@ use lang\{Closeable, Value};
  * @deprecated by XmlStreaming
  * @test  util.address.unittest.XmlInputTest
  */
-class XmlFile extends Address implements Closeable, Value {
-  private $file;
+class XmlFile extends Address {
 
   /**
    * Creates a new file-based XML input
@@ -18,34 +17,10 @@ class XmlFile extends Address implements Closeable, Value {
    * @param  string|io.Path|io.File $file
    */
   public function __construct($file) {
-    $this->file= $file instanceof File ? $file : new File($file);
+    parent::__construct(new FileInputStream($file));
   }
 
-  /** @return io.streams.InputStream */
-  protected function stream() { return $this->file->in(); }
+  /** Iterator implementation */
+  protected function iterator(): Iterator { return new XmlIterator($this->stream); }
 
-  /** @return string */
-  public function toString() {
-    return nameof($this).'<'.$this->file->toString().'>';
-  }
-
-  /** @return string */
-  public function hashCode() {
-    return 'F'.$this->file->hashCode();
-  }
-
-  /**
-   * Comparison
-   *
-   * @param  var $value
-   * @return int
-   */
-  public function compareTo($value) {
-    return $value instanceof self ? $this->file->compareTo($value->file) : 1;
-  }
-
-  /** @return void */
-  public function close() {
-    $this->file->isOpen() && $this->file->close();
-  }
 }
