+++ src/main/php/util/address/XmlStream.class.php
@@ -1,8 +1,6 @@
 <?php namespace util\address;
 
-use io\streams\InputStream;
-use lang\{Closeable, Value};
-use util\Objects;
+use Iterator;
 
 /**
  * XML stream input
@@ -11,50 +9,9 @@ use util\Objects;
  * @test  util.address.unittest.XmlInputTest
  * @deprecated by XmlStreaming
  */
-class XmlStream extends Address implements Closeable, Value {
-  private $in;
+class XmlStream extends Address {
 
-  /**
-   * Creates a new stream-based XML input
-   *
-   * @param  io.streams.InputStream $in
-   */
-  public function __construct(InputStream $in) {
-    $this->in= $in;
-  }
+  /** Iterator implementation */
+  protected function iterator(): Iterator { return new XmlIterator($this->stream); }
 
-  /** @return io.streams.InputStream */
-  protected function stream() { return $this->in; }
-
-  /** @return string */
-  public function toString() {
-
-    // Most InputStream instances have a toString() method but do not implement
-    // the Value interface, see https://github.com/xp-framework/core/issues/310
-    if ($this->in instanceof Value || method_exists($this->in, 'toString')) {
-      return nameof($this).'<'.$this->in->toString().'>';
-    } else {
-      return nameof($this).'<'.Objects::stringOf($this->in).'>';
-    }
-  }
-
-  /** @return string */
-  public function hashCode() {
-    return 'S'.Objects::hashOf($this->in);
-  }
-
-  /**
-   * Comparison
-   *
-   * @param  var $value
-   * @return int
-   */
-  public function compareTo($value) {
-    return $value instanceof self ? Objects::compare($this->in, $value->in) : 1;
-  }
-
-  /** @return void */
-  public function close() {
-    $this->in->close();
-  }
 }
