+++ src/main/php/util/address/XmlString.class.php
@@ -1,7 +1,7 @@
 <?php namespace util\address;
 
+use Iterator;
 use io\streams\MemoryInputStream;
-use lang\Value;
 
 /**
  * XML string input
@@ -9,8 +9,7 @@ use lang\Value;
  * @test  util.address.unittest.XmlInputTest
  * @deprecated by XmlStreaming
  */
-class XmlString extends Address implements Value {
-  private $input;
+class XmlString extends Address {
 
   /**
    * Creates a new file-based XML input
@@ -18,30 +17,9 @@ class XmlString extends Address implements Value {
    * @param  string $input
    */
   public function __construct($input) {
-    $this->input= $input;
+    parent::__construct(new MemoryInputStream($input));
   }
 
-  /** @return io.streams.InputStream */
-  protected function stream() { return new MemoryInputStream($this->input); }
-
-  /** @return string */
-  public function toString() {
-    $l= strlen($this->input);
-    return nameof($this).'<"'.($l > 10 ? substr($this->input, 0, 10).'..." ('.$l.' bytes)>' : $this->input.'">');
-  }
-
-  /** @return string */
-  public function hashCode() {
-    return md5($this->input);
-  }
-
-  /**
-   * Comparison
-   *
-   * @param  var $value
-   * @return int
-   */
-  public function compareTo($value) {
-    return $value instanceof self ? $this->input <=> $value->input : 1;
-  }
+  /** Iterator implementation */
+  protected function iterator(): Iterator { return new XmlIterator($this->stream); }
 }
